requirements.txt
pdfplumber>=0.11.0
pandas>=2.0.0
matplotlib>=3.7.0
pytrends>=4.9.2
pyyaml>=6.0.0
nltk>=3.8.0
streamlit>=1.31.0
huggingface-hub>=0.20.0
